Fix incorrect safety annotations
The `Tls` and `TlsConfig` functions require, as an invariant,
that the wrapped pointer does not dangle. This invariant must be
maintained by all available constructors for these objects.

The `From` implementations did not maintain it. They checked for null
pointers, but they did not and cannot check for dangling pointers. They
were also marked as safe. This allows one to use rust-libtls to invoke
undefined behavior in safe code.

    // Safe Rust allows you to construct arbitrary raw pointers.
    let my_raw_ptr = 3 as usize as *mut libtls_sys::tls;
    // Violate the invariant that Tls wraps a valid tls pointer.
    let my_tls_object = Tls::from(my_raw_ptr);
    // Invoke undefined behavior by dereferencing it in tls_reset.
    my_tls_object.reset();